# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Legend Bucks rewards platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: managerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get an employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
  /employees/{id}/balance:
    get:
      operationId: getEmployeeBalance
      tags: [employees]
      summary: Get an employee's current balance calculated from ledger
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceSummary"
  /employees/{id}/deactivate:
    patch:
      operationId: deactivateEmployee
      tags: [employees]
      summary: Deactivate an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"

  # Budgets
...
  /dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /dashboard/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [dashboard]
      summary: Top employees by bucks received this month
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
components:
  schemas:
...
    Employee:
      type: object
      required: [id, firstName, lastName, email, role, status, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        department:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        managerId:
          type: ["integer", "null"]
        managerName:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, manager, team_member]
        status:
          type: string
          enum: [active, inactive, invited]
        balance:
          type: ["integer", "null"]
        createdAt:
          type: string
    EmployeeUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        department:
          type: string
        location:
          type: string
        managerId:
          type: ["integer", "null"]
        role:
          type: string
          enum: [admin, manager, team_member]
        status:
          type: string
          enum: [active, inactive]
    BalanceSummary:
      type: object
      required: [employeeId, balance, totalReceived, totalSpent]
      properties:
        employeeId:
          type: integer
        balance:
          type: integer
        totalReceived:
          type: integer
        totalSpent:
          type: integer
...
    ActivityItem:
      type: object
      required: [id, type, description, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [award, redemption, goal_contribution, new_employee, budget_assigned]
        description:
          type: string
        actorName:
          type: ["string", "null"]
        targetName:
          type: ["string", "null"]
        amount:
          type: ["integer", "null"]
        createdAt:
          type: string
    LeaderboardEntry:
      type: object
      required: [rank, employeeId, employeeName, bucksReceived]
      properties:
        rank:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        department:
          type: ["string", "null"]
        bucksReceived:
          type: integer