# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Legend Bucks rewards platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemType } from './activityItemType';

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  /** @nullable */
  actorName?: string | null;
  /** @nullable */
  targetName?: string | null;
  /** @nullable */
  amount?: number | null;
  createdAt: string;
}